Move inferDomainFromFile
